feat : added functionality of quantity in wishlist
